Make Harbor runtime portable
