+++ package.json
@@ -10,6 +10,7 @@
     "lint": "tsc -b --pretty false",
     "clean": "pnpm -r clean",
     "package:agent-cli": "pnpm build && node scripts/package-agent-cli.mjs",
+    "package:harbor-runtime": "node scripts/package-harbor-runtime.mjs",
     "smoke:local": "pnpm build && node scripts/smoke-local.mjs",
     "smoke:local:fake": "pnpm build && node scripts/smoke-local.mjs --provider fake",
     "smoke:harbor": "node scripts/run-bash.mjs scripts/smoke-harbor.sh",
@@ -20,6 +21,7 @@
     "bench:tb:deepseek:k5": "node scripts/bench-terminal-bench.mjs --mode k --provider deepseek --model deepseek-v4-pro --k 5",
     "bench:tb:deepseek:k10": "node scripts/bench-terminal-bench.mjs --mode k --provider deepseek --model deepseek-v4-pro --k 10",
     "bench:tb:deepseek:task": "node scripts/bench-terminal-bench.mjs --mode task --provider deepseek --model deepseek-v4-pro",
+    "bench:tb:config:portable": "node scripts/package-harbor-runtime.mjs",
     "bench:tb:report": "node scripts/bench-report.mjs",
     "bench:tb:report:all": "node scripts/bench-report.mjs --all"
   },
